requirements.txt
annotated-types==0.7.0
anyio==4.10.0
Authlib==1.6.2
boto3==1.40.16
botocore==1.40.16
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
cryptography==45.0.6
exceptiongroup==1.3.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
jmespath==1.0.1
packaging==25.0
playwright==1.54.0
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.1
pytest-base-url==2.1.0
pytest-playwright==0.7.0
pytest-xprocess==0.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
requests==2.32.5
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
text-unidecode==1.3
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.20
uvicorn==0.35.0
